+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Support the `hasPragma` function.
+- Support the new `number_arg` node type in Ruby 2.7.
 
 ## [0.8.0] - 2019-03-08
 ### Added